requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Streamlit
streamlit==1.31.0

# Image Processing
Pillow==10.2.0
pdf2image==1.17.0

# AI/ML
litellm==1.25.0
openai==1.12.0

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
aiofiles==23.2.1

# Logging
loguru==0.7.2

# Testing (optional)
pytest==8.0.0
httpx==0.26.0
